backend/requirements.txt
fastapi
uvicorn[standard]
transformers
sentence-transformers
langchain
langchain-community
langchain-core
faiss-cpu
psycopg2-binary
chromadb
accelerate
unstructured
pandas
python-multipart
llama-cpp-python
langchain-experimental
psycopg2-binary
python-dotenv